3.x compilation and test fixes
Deleted some #include "buffer.h" lines - buffer modules have been removed for now.

SQL Server unit tests needed to default verbosity option - otherwise it defaulted to None and
was compared against a number by the unit test framework.  Not allowed in Python 3 (which is
good - it was incorrect but I didn't notice in 2).